Удалил никогда по факту не работавший метод GetProtocolVersion и переделал старт сессии
