+++ src/OneScript.DebugServices/DefaultDebugService.cs
@@ -22,8 +22,6 @@ namespace OneScript.DebugServices
 {
     public class DefaultDebugService : IDebuggerService
     {
-        private const int CURRENT_PROTOCOL_VERSION = 3;
-        
         private readonly IBreakpointManager _breakpointManager;
         private readonly IVariableVisualizer _visualizer;
         private readonly ThreadManager _threadManager;
@@ -248,11 +246,6 @@ namespace OneScript.DebugServices
         private IList<IVariable> GetChildVariables(IVariable src)
         {
             return _visualizer.GetChildVariables(src.Value).ToList();
-        }
-
-        public int GetProtocolVersion()
-        {
-            return CURRENT_PROTOCOL_VERSION;
         }
     }
 }
