+++ src/VSCode.DebugAdapter/DebugeeProcess.cs
@@ -13,7 +13,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using Newtonsoft.Json.Linq;
 using Serilog;
-using VSCode.DebugAdapter.OscriptProtocols;
+using VSCode.DebugAdapter.Transport;
 using StackFrame = OneScript.DebugProtocol.StackFrame;
 
 namespace VSCode.DebugAdapter
@@ -29,7 +29,7 @@ namespace VSCode.DebugAdapter
         
         private Encoding _dapEncoding;
 
-        private TcpDebugServerClient _debugger;
+        private OneScriptDebuggerClient _debugger;
 
         private readonly PathHandlingStrategy _strategy;
 
@@ -42,8 +42,6 @@ namespace VSCode.DebugAdapter
             _strategy = pathHandling;
         }
         
-        public string DebugProtocol { get; protected set; }
-        
         public bool HasExited => _process?.HasExited ?? true;
         public int ExitCode => _process.ExitCode;
 
@@ -54,8 +52,8 @@ namespace VSCode.DebugAdapter
             get => _activeProtocolVersion;
             set
             {
+                ValidateProtocolVersion(value);
                 _activeProtocolVersion = value;
-                SetupSupportedProtocolVersion();
             }
         }
 
@@ -137,9 +135,10 @@ namespace VSCode.DebugAdapter
             return RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? null : Encoding.UTF8;
         }
 
-        public void SetConnection(TcpDebugServerClient service)
+        public void SetClient(OneScriptDebuggerClient service)
         {
             _debugger = service;
+            ProtocolVersion = service.ProtocolVersion;
         }
         
         public event EventHandler<DebugeeOutputEventArgs> OutputReceived;
@@ -147,7 +146,7 @@ namespace VSCode.DebugAdapter
         
         private void Process_Exited(object sender, EventArgs e)
         {
-            _debugger?.Disconnect();
+            _debugger?.Stop();
             Terminate();
             ProcessExited?.Invoke(this, new EventArgs());
         }
@@ -161,29 +160,11 @@ namespace VSCode.DebugAdapter
             RaiseOutputReceivedEvent("stdout", e.Data);
         }
 
-        private void SetupSupportedProtocolVersion()
+        private void ValidateProtocolVersion(int value)
         {
-            if (!ProtocolVersions.IsValid(_activeProtocolVersion))
+            if (!ProtocolVersions.IsValid(value))
             {
-                _activeProtocolVersion = ProtocolVersions.SafestVersion;
-                return;
-            }
-
-            if (_activeProtocolVersion != ProtocolVersions.UnknownVersion)
-            {
-                // Задали вручную корректное значение. Ничего не запрашиваем
-                return;
-            }
-
-            try
-            {
-                Log.Debug("Requesting max protocol version");
-                _activeProtocolVersion = _debugger.GetProtocolVersion();
-                Log.Information("Protocol version is {ProtocolVersion}", _activeProtocolVersion);
-            }
-            catch (Exception e)
-            {
-                Log.Error(e, "Unknown error while checking version");
+                throw new ArgumentOutOfRangeException($"Protocol version {value} is unknown.");
             }
         }
