+++ src/VSCode.DebugAdapter/Transport/TransportException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace VSCode.DebugAdapter.Transport
+{
+    public class TransportException : ApplicationException
+    {
+        public TransportException(string message) : base(message)
+        {
+        }
+    }
+}
