+++ src/VSCode.DebugAdapter/Transport/TransportProtocols.cs
@@ -1,0 +1,10 @@
+namespace VSCode.DebugAdapter.Transport
+{
+    public enum TransportProtocols
+    {
+        Auto,
+        Binary,
+        Json,
+        Invalid,
+    }
+}
