+++ src/VSCode.DebugAdapter/package.json
@@ -113,11 +113,6 @@
                 "type": "string",
                 "description": "Кодировка вывода отлаживаемого приложения. Отладчик будет интерпретировать вывод приложения в указанной кодировке",
                 "default": ""
-              },
-              "protocolVersion": {
-                "type": "number",
-                "description": "Версия протокола отладки, поддерживаемая движком 1Скрипт",
-                "default": 0
               }
             }
           },
@@ -132,11 +127,6 @@
                 "type": "string",
                 "description": "Кодировка вывода отлаживаемого приложения. Отладчик будет интерпретировать вывод приложения в указанной кодировке",
                 "default": ""
-              },
-              "protocolVersion": {
-                "type": "number",
-                "description": "Версия протокола отладки, поддерживаемая движком 1Скрипт",
-                "default": 0
               }
             }
           }
